Add logging to pandocomatic (#112); prepare 2.0
Pandocomatic now logs what it is doing to file. When using detail level
"debug", it logs every step in the conversion process. This includes the
final template used for the conversion, all processors that are run, and
the full pandoc invocation.

To implement this feature, I made the following changes:

- Remove command-line option `--debug` (breaking change). Instead, use
  command-line options `--log` and `--log-level debug`.

- Add command-line option `--log [FILE]`. This tells pandocomatic to log
  its doings to `FILE`. If you don't supply a file, pandocomatic logs to
  "pandocomatic.log" by default.

- Add command-line option `--log-level [LEVEL]`. This tells pandocomatic
  what level of detail to log. Possible values are: "fatal", "error",
  "warning", "info", and "debug". Default value is "info".

  This option only has an effect when used together with command-line
  option `--log`.

  Pick level "debug" for a complete overview of the conversion process.

Because these changes are a breaking change in pandocomatic's
interface, I am preparing a new major version: 2.0. To that end, I'm
also updating dependencies and the minimal ruby version to 3.2.x.